add demo project

+++ shared/src/types.rs
@@ -1,1 +1,1 @@
-pub type EntityId = u32;
+pub type EntityId = u64;
